Move procedure-not-in-module to same file as external-procedure
Renumbers from C111 to C092
