+++ fortitude/resources/test/fixtures/correctness/C092.f90
